execution-engine/requirements.txt
aio-pika==9.3.1
aiormq==6.7.7
alembic==1.12.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.2.5
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
pamqp==3.2.1
passlib==1.7.4
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0.post1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
